refactor(types): migrate types
