+++ src/components/ExportData.vue
@@ -45,7 +45,6 @@ import { Getter } from 'vuex-class'
 import { ipcRenderer } from 'electron'
 import { loadConnections } from '@/api/connection'
 import MyDialog from './MyDialog.vue'
-import { ConnectionModel } from '@/views/connections/types'
 import XMLConvert from 'xml-js'
 import { parse as CSVConvert } from 'json2csv'
 import ExcelConvert, { WorkBook } from 'xlsx'
