+++ src/components/ImportData.vue
@@ -60,7 +60,6 @@ import { Getter } from 'vuex-class'
 import fs from 'fs'
 import { remote } from 'electron'
 import { importConnections } from '@/api/connection'
-import { ConnectionModel } from '@/views/connections/types'
 import MyDialog from './MyDialog.vue'
 import XMLConvert from 'xml-js'
 import CSVConvert from 'csvtojson'
