+++ src/components/SubscriptionsList.vue
@@ -112,7 +112,6 @@ import { updateConnection } from '@/api/connection'
 import { defineColors, getRandomColor } from '@/utils/colors'
 import LeftPanel from '@/components/LeftPanel.vue'
 import MyDialog from '@/components/MyDialog.vue'
-import { ConnectionModel, SubscribeErrorReason } from '../views/connections/types'
 import VueI18n from 'vue-i18n'
 
 @Component({
