+++ src/views/connections/ConnectionForm.vue
@@ -480,7 +480,6 @@ import { Component, Vue, Prop, Watch } from 'vue-property-decorator'
 import { Getter, Action } from 'vuex-class'
 import { loadConnection, createConnection, updateConnection, loadSuggestConnections } from '@/api/connection'
 import getClientId from '@/utils/getClientId'
-import { ConnectionModel, SearchCallBack, NameCallBack, FormRule } from './types'
 import { getMQTTProtocol } from '@/utils/mqttUtils'
 import Editor from '@/components/Editor.vue'
 import deepMerge from '@/utils/deepMerge'
