+++ src/views/connections/index.vue
@@ -33,7 +33,6 @@ import EmptyPage from '@/components/EmptyPage.vue'
 import ConnectionsList from './ConnectionsList.vue'
 import ConnectionsDetail from './ConnectionsDetail.vue'
 import ConnectionForm from './ConnectionForm.vue'
-import { ConnectionModel, ConnectionModelCollection } from './types'
 
 @Component({
   components: {
